Reworked Decode module structure
With the original Decode.Internal moved to Decode.Parsers, created
a new Decode.Internal which is now exposed, and moved some functions
there from Decode which are only needed for testing, and can't
reliably be used except as part of decoding a full message:

    - decodeDNSHeader
    - decodeDNSFlags
    - decodeDomain
    - decodeMailbox
    - decodeResourceRecord
    - decodeResourceRecordAt

+++ Network/DNS/IO.hs
@@ -49,7 +49,7 @@ receive sock = do
     let bufsiz = fromIntegral maxUdpSize
     bs <- recv sock bufsiz `E.catch` \e -> E.throwIO $ NetworkFailure e
     Elapsed (Seconds now) <- timeCurrent
-    case decodeAt (fromIntegral now) bs of
+    case decodeAt now bs of
         Left  e   -> E.throwIO e
         Right msg -> return msg
 
@@ -62,7 +62,7 @@ receiveVC sock = do
     len <- toLen <$> recvDNS sock 2
     bs <- recvDNS sock len
     Elapsed (Seconds now) <- timeCurrent
-    case decodeAt (fromIntegral now) bs of
+    case decodeAt now bs of
         Left e    -> E.throwIO e
         Right msg -> return msg
   where
